First batch

+++ libc/src/ctype/tolower.cpp
@@ -16,7 +16,7 @@ namespace __llvm_libc {
 // TODO: Currently restricted to default locale.
 // These should be extended using locale information.
 LLVM_LIBC_FUNCTION(int, tolower, (int c)) {
-  if (internal::isupper(c))
+  if (internal::isupper(static_cast<unsigned>(c)))
     return c + 'a' - 'A';
   return c;
 }
